# docs/architecture-pack/contracts/openapi.yaml
openapi: "3.1.0"
info:
  title: Revizr API
  version: "1.0.0"
  description: |
    Revizr backend API — exam revision platform for UK students (KS3, GCSE, A-Level, 11+).
    All endpoints require Bearer JWT authentication unless noted as public.
    Base URL: https://api.revizr.co.uk/v1
  contact:
    name: Revizr Engineering
    email: engineering@revizr.co.uk
  license:
    name: Proprietary
servers:
  - url: https://api.revizr.co.uk/v1
    description: Production (eu-west-2)
  - url: https://api.staging.revizr.co.uk/v1
    description: Staging (eu-west-2)
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Access token issued by POST /auth/login or POST /auth/refresh.
        Expires in 15 minutes. Pass in Authorization header as "Bearer <token>".
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "VALIDATION_ERROR"
        message:
          type: string
          example: "email is required"
        details:
          type: object
          example: { "field": "email", "constraint": "format" }
...
  responses:
    Unauthorized:
      description: Missing or invalid access token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired access token"
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: "FORBIDDEN"
            message: "You do not have permission to access this resource"
...
paths:
...
  /accounts/pair:
    post:
      tags: [accounts]
      operationId: accountsPair
      summary: Initiate parent–student account pairing
      description: >
        Parent initiates pairing by providing student's email. Sends a
        consent/pairing invitation to the student's registered email.
        Requires the calling user to have role=parent.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [student_email, relationship]
              properties:
                student_email:
                  type: string
                  format: email
                  example: "amara@example.com"
                relationship:
                  type: string
                  enum: [parent, guardian, carer]
                  example: "parent"
      responses:
        "202":
          description: Pairing invitation sent
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: string
                    format: uuid
                    example: "aa11bb22-cc33-dd44-ee55-ff6677889900"
                  status:
                    type: string
                    example: "pending"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Student account not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "STUDENT_NOT_FOUND"
                message: "No account found with this email"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /accounts/consent:
    post:
      tags: [accounts]
      operationId: accountsConsent
      summary: Record parental consent for a paired account
      description: >
        Called when a parent confirms consent via email link or in-app flow.
        Records consent timestamp, mechanism, and IP hash. Activates the pairing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [account_id, consent_token]
              properties:
                account_id:
                  type: string
                  format: uuid
                  example: "aa11bb22-cc33-dd44-ee55-ff6677889900"
                consent_token:
                  type: string
                  description: Token from the consent email link
                  example: "b2c3d4e5f6a7890123456789012345678901bcde"
      responses:
        "200":
          description: Consent recorded; account pair activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  account_id:
                    type: string
                    format: uuid
                    example: "aa11bb22-cc33-dd44-ee55-ff6677889900"
                  status:
                    type: string
                    example: "active"
        "400":
          description: Invalid or expired consent token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: "INVALID_CONSENT_TOKEN"
                message: "Consent token is invalid or has expired"

  # ─── USERS ───────────────────────────────────────────────────────────────────
...